Added configuration system backend to knoxite
This commits adds a configuration system backend to knoxite.
The code is heavily inspired by the configuration backend for beehive written
by rubiojr so all your credits are belong to him:
https://github.com/muesli/beehive/commit/5f810aba2ea036934b43c57a922b307afdaeff2f

The main difference is ofc the adjustments to work with/for knoxite.
I also modified the password system as I'm not feeling safe setting the password
for the config file in an environment variable. It'll prompt the user once when
there's no password specified in the url.